Almost max out what can be done in float64 without losing speed: Also data in float64 (not float32) to avoid precision loss in initial operations

+++ experiments/tagging/experiment.py
@@ -261,7 +261,7 @@ class TaggingExperiment(BaseExperiment):
     def _get_ypred_and_label(self, batch):
         batch = batch.to(self.device)
         embedding = embed_tagging_data(
-            batch.x.to(self.dtype),
+            batch.x,
             batch.scalars.to(self.dtype),
             batch.ptr,
             self.cfg.data,
